# src/main/resources/openapi/openapi.yaml
openapi: 3.0.3
info:
  title: MOCA Backend API
  description: |
    결제 전 최적 카드 추천을 위한 MOCA Backend REST API입니다.

    Google PKCE 로그인은 PWA가 authorization code를 받은 뒤 MOCA 서버가 Client Secret으로 교환합니다.
    이후 MOCA API는 서버가 발급한 opaque access token으로 인증합니다.
  version: 1.0.0
servers:
  - url: /api/v1
    description: MOCA API v1
...
paths:
...
  /benefit-history:
    get:
      tags: [Benefit]
      summary: 혜택 내역 목록 조회
      description: yearMonth를 생략하면 Asia/Seoul 기준 현재 월을 사용합니다. 확정된 할인·캐시백·포인트 혜택만 반환합니다.
      operationId: getBenefitHistory
      security:
        - MocaBearerAuth: []
      parameters:
        - $ref: '#/components/parameters/YearMonth'
        - name: userCardId
          in: query
          schema: { type: string, format: uuid }
        - name: type
          in: query
          schema: { type: string, enum: [DISCOUNT, CASHBACK, POINT] }
        - name: sort
          in: query
          schema: { type: string, enum: [LATEST, BENEFIT_DESC], default: LATEST }
        - name: page
          in: query
          schema: { type: integer, minimum: 1, default: 1 }
        - name: size
          in: query
          schema: { type: integer, minimum: 1, maximum: 100, default: 20 }
      responses:
        '200':
          description: 혜택 이력 목록
          content:
            application/json:
              schema: { $ref: '#/components/schemas/BenefitHistoryApiResponse' }
        '400': { $ref: '#/components/responses/ValidationError' }
        '401': { $ref: '#/components/responses/AuthenticationError' }
  /benefit-history/{benefitHistoryId}:
    get:
      tags: [Benefit]
      summary: 혜택 내역 상세 조회
      operationId: getBenefitHistoryDetail
      security:
        - MocaBearerAuth: []
      parameters:
        - name: benefitHistoryId
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        '200':
          description: 혜택 이력 상세
          content:
            application/json:
              schema: { $ref: '#/components/schemas/BenefitHistoryDetailApiResponse' }
        '401': { $ref: '#/components/responses/AuthenticationError' }
        '404':
          description: 로그인 사용자의 혜택 이력이 없음
  /reports/benefits/summary:
    get:
      tags: [Report]
      summary: 월별 혜택 합계와 전월 비교 조회
      operationId: getBenefitReportSummary
      security: [{ MocaBearerAuth: [] }]
      parameters: [{ $ref: '#/components/parameters/YearMonth' }]
      responses:
        '200':
          description: 월별 혜택 합계
          content:
            application/json:
              schema: { $ref: '#/components/schemas/BenefitSummaryReportApiResponse' }
        '400': { $ref: '#/components/responses/ValidationError' }
        '401': { $ref: '#/components/responses/AuthenticationError' }
  /reports/benefits/categories:
    get:
      tags: [Report]
      summary: 월별 혜택 상위 가맹점 카테고리 조회
      description: merchant가 매핑된 승인내역의 확정 혜택만 집계합니다.
      operationId: getBenefitReportCategories
      security: [{ MocaBearerAuth: [] }]
      parameters:
        - $ref: '#/components/parameters/YearMonth'
        - name: limit
          in: query
          schema: { type: integer, minimum: 1, maximum: 3, default: 3 }
      responses:
        '200':
          description: 월별 상위 카테고리
          content:
            application/json:
              schema: { $ref: '#/components/schemas/BenefitCategoriesReportApiResponse' }
        '400': { $ref: '#/components/responses/ValidationError' }
        '401': { $ref: '#/components/responses/AuthenticationError' }
  /reports/benefits/missed:
    get:
      tags: [Report]
      summary: 카드별 실제 미적용 혜택 조회
      description: 승인별 계산 결과의 예상 혜택과 실제 적용 혜택 차이만 반환합니다. 단순 월 한도 잔여분은 포함하지 않습니다.
      operationId: getBenefitReportMissed
      security: [{ MocaBearerAuth: [] }]
      parameters:
        - $ref: '#/components/parameters/YearMonth'
        - name: userCardId
          in: query
          required: true
          schema: { type: string, format: uuid }
      responses:
        '200':
          description: 실제 미적용 혜택
          content:
            application/json:
              schema: { $ref: '#/components/schemas/MissedBenefitsReportApiResponse' }
        '400': { $ref: '#/components/responses/ValidationError' }
        '401': { $ref: '#/components/responses/AuthenticationError' }
        '404': { description: 본인 소유의 활성 카드를 찾을 수 없음 }
  /reports/performances/summary:
    get:
      tags: [Report]
      summary: 월 실적 도넛과 대표 카드 조회
      operationId: getPerformanceReportSummary
      security: [{ MocaBearerAuth: [] }]
      parameters: [{ $ref: '#/components/parameters/YearMonth' }]
      responses:
        '200':
          description: 실적 요약
          content:
            application/json:
              schema: { $ref: '#/components/schemas/PerformanceSummaryReportApiResponse' }
        '400': { $ref: '#/components/responses/ValidationError' }
        '401': { $ref: '#/components/responses/AuthenticationError' }
  /reports/performances/cards:
    get:
      tags: [Report]
      summary: 월별 전체 카드 실적 조회
      description: 현재 카드 원천 데이터에는 실적 tier가 없어 currentTier=0, nextTier=null로 반환합니다.
      operationId: getPerformanceReportCards
      security: [{ MocaBearerAuth: [] }]
      parameters: [{ $ref: '#/components/parameters/YearMonth' }]
      responses:
        '200':
          description: 카드별 실적 현황
          content:
            application/json:
              schema: { $ref: '#/components/schemas/PerformanceCardsReportApiResponse' }
        '400': { $ref: '#/components/responses/ValidationError' }
        '401': { $ref: '#/components/responses/AuthenticationError' }
...
components:
  securitySchemes:
    MocaBearerAuth:
      type: http
      scheme: bearer
      bearerFormat: opaque
      description: Google token이 아닌 MOCA 서버 발급 opaque access token을 입력합니다.
  parameters:
    YearMonth:
      name: yearMonth
      in: query
      required: false
      description: 집계 기준 월(미전달 시 Asia/Seoul 현재 월)
      schema:
        type: string
        pattern: '^\d{4}-(0[1-9]|1[0-2])$'
        example: 2026-07
  responses:
    AuthenticationError:
      description: access token이 없거나 유효하지 않음
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
    ValidationError:
      description: 요청값 검증 실패
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
  schemas:
...
    BenefitHistoryApiResponse:
      type: object
      properties:
        success: { type: boolean, example: true }
        data: { $ref: '#/components/schemas/BenefitHistoryResponse' }
    BenefitHistoryResponse:
      type: object
      required: [data, meta]
      properties:
        data:
          type: array
          items: { $ref: '#/components/schemas/BenefitHistoryItemResponse' }
        meta: { $ref: '#/components/schemas/BenefitHistoryMetaResponse' }
    BenefitHistoryItemResponse:
      type: object
      required: [benefitHistoryId, merchantName, approvedAt, paymentAmount, benefitAmount, benefitType, benefitTitle, userCardId, cardName, calculationStatus]
      properties:
        benefitHistoryId: { type: string, format: uuid }
        merchantName: { type: string }
        approvedAt: { type: string, format: date-time }
        paymentAmount: { type: integer, format: int64 }
        benefitAmount: { type: integer, format: int64 }
        benefitType: { type: string, enum: [DISCOUNT, CASHBACK, POINT] }
        benefitTitle: { type: string }
        userCardId: { type: string, format: uuid }
        cardName: { type: string }
        calculationStatus: { type: string, enum: [APPLIED] }
    BenefitHistoryMetaResponse:
      type: object
      required: [page, size, totalCount, hasNext]
      properties:
        page: { type: integer }
        size: { type: integer }
        totalCount: { type: integer, format: int64 }
        hasNext: { type: boolean }
    BenefitHistoryDetailApiResponse:
      type: object
      properties:
        success: { type: boolean, example: true }
        data: { $ref: '#/components/schemas/BenefitHistoryDetailResponse' }
    BenefitHistoryDetailResponse:
      type: object
      required: [benefitHistoryId, calculationStatus, merchantName, approvedAt, cardName, paymentAmount, benefitAmount, benefitType, benefitTitle, monthlyLimit]
      properties:
        benefitHistoryId: { type: string, format: uuid }
        calculationStatus: { type: string, enum: [APPLIED] }
        merchantName: { type: string }
        approvedAt: { type: string, format: date-time }
        cardName: { type: string }
        paymentAmount: { type: integer, format: int64 }
        benefitAmount: { type: integer, format: int64 }
        benefitType: { type: string, enum: [DISCOUNT, CASHBACK, POINT] }
        benefitTitle: { type: string }
        monthlyLimit: { $ref: '#/components/schemas/MonthlyLimitResponse' }
        earnedMileage: { type: integer, format: int64, nullable: true, description: 마일리지 적립 건에서만 반환 }
    MonthlyLimitResponse:
      type: object
      required: [usedAmount, limitAmount, remainingAmount]
      properties:
        usedAmount: { type: integer, format: int64 }
        limitAmount: { type: integer, format: int64 }
        remainingAmount: { type: integer, format: int64 }
    BenefitSummaryReportApiResponse:
      type: object
      required: [success, data]
      properties:
        success: { type: boolean, example: true }
        data: { $ref: '#/components/schemas/BenefitSummaryReportResponse' }
    BenefitSummaryReportResponse:
      type: object
      required: [yearMonth, totalBenefitAmount, previousMonthBenefitAmount, differenceAmount, breakdown]
      properties:
        yearMonth: { type: string, pattern: '^\\d{4}-(0[1-9]|1[0-2])$' }
        totalBenefitAmount: { type: integer, format: int64, description: 원화 가치가 있는 확정 혜택 합계 }
        previousMonthBenefitAmount: { type: integer, format: int64 }
        differenceAmount: { type: integer, format: int64 }
        breakdown:
          type: array
          items: { $ref: '#/components/schemas/BenefitBreakdownResponse' }
    BenefitBreakdownResponse:
      type: object
      required: [type, label, amount]
      properties:
        type: { type: string, enum: [DISCOUNT, CASHBACK, POINT] }
        label: { type: string, example: 할인 }
        amount: { type: integer, format: int64 }
    BenefitCategoriesReportApiResponse:
      type: object
      required: [success, data]
      properties:
        success: { type: boolean, example: true }
        data: { $ref: '#/components/schemas/BenefitCategoriesReportResponse' }
    BenefitCategoriesReportResponse:
      type: object
      required: [yearMonth, categories]
      properties:
        yearMonth: { type: string, pattern: '^\\d{4}-(0[1-9]|1[0-2])$' }
        categories:
          type: array
          maxItems: 3
          items: { $ref: '#/components/schemas/BenefitCategoryResponse' }
    BenefitCategoryResponse:
      type: object
      required: [rank, categoryCode, categoryName, benefitAmount]
      properties:
        rank: { type: integer, minimum: 1 }
        categoryCode: { type: string, example: CAFE }
        categoryName: { type: string, example: 카페 }
        benefitAmount: { type: integer, format: int64 }
    MissedBenefitsReportApiResponse:
      type: object
      required: [success, data]
      properties:
        success: { type: boolean, example: true }
        data: { $ref: '#/components/schemas/MissedBenefitsReportResponse' }
    MissedBenefitsReportResponse:
      type: object
      required: [yearMonth, userCard, totalMissedBenefitAmount, benefits]
      properties:
        yearMonth: { type: string, pattern: '^\\d{4}-(0[1-9]|1[0-2])$' }
        userCard: { $ref: '#/components/schemas/ReportUserCardResponse' }
        totalMissedBenefitAmount:
          type: integer
          format: int64
          description: 승인별 계산 결과에서 확인된 실제 미적용 혜택 합계
        benefits:
          type: array
          items: { $ref: '#/components/schemas/MissedBenefitItemResponse' }
    ReportUserCardResponse:
      type: object
      required: [userCardId, cardName]
      properties:
        userCardId: { type: string, format: uuid }
        cardName: { type: string }
        cardColor: { type: string, nullable: true, description: 현재 카드 카탈로그에 색상 원천이 없어 null 가능 }
    MissedBenefitItemResponse:
      type: object
      required: [benefitRuleId, title, type, usedAmount, limitAmount, remainingAmount, unit]
      properties:
        benefitRuleId: { type: string, format: uuid }
        title: { type: string }
        type: { type: string, enum: [DISCOUNT, CASHBACK, POINT] }
        usedAmount: { type: integer, format: int64 }
        limitAmount: { type: integer, format: int64 }
        remainingAmount: { type: integer, format: int64 }
        unit: { type: string, enum: [KRW] }
    PerformanceSummaryReportApiResponse:
      type: object
      required: [success, data]
      properties:
        success: { type: boolean, example: true }
        data: { $ref: '#/components/schemas/PerformanceSummaryReportResponse' }
    PerformanceSummaryReportResponse:
      type: object
      required: [yearMonth, cardCount, achievedCardCount, cards]
      properties:
        yearMonth: { type: string, pattern: '^\\d{4}-(0[1-9]|1[0-2])$' }
        cardCount: { type: integer }
        achievedCardCount: { type: integer }
        cards:
          type: array
          maxItems: 3
          items: { $ref: '#/components/schemas/PerformanceSummaryCardResponse' }
    PerformanceSummaryCardResponse:
      type: object
      required: [userCardId, cardName, achievementRate, isCurrentTierAchieved]
      properties:
        userCardId: { type: string, format: uuid }
        cardName: { type: string }
        achievementRate: { type: integer, minimum: 0, maximum: 100 }
        isCurrentTierAchieved: { type: boolean }
    PerformanceCardsReportApiResponse:
      type: object
      required: [success, data]
      properties:
        success: { type: boolean, example: true }
        data: { $ref: '#/components/schemas/PerformanceCardsReportResponse' }
    PerformanceCardsReportResponse:
      type: object
      required: [yearMonth, cards]
      properties:
        yearMonth: { type: string, pattern: '^\\d{4}-(0[1-9]|1[0-2])$' }
        cards:
          type: array
          items: { $ref: '#/components/schemas/PerformanceCardResponse' }
    PerformanceCardResponse:
      type: object
      required: [userCardId, cardName, currentPerformanceAmount, currentTierTargetAmount,
                 achievementRate, currentTier, isCurrentTierAchieved, remainingAmountToNextTier]
      properties:
        userCardId: { type: string, format: uuid }
        cardName: { type: string }
        cardImageUrl: { type: string, format: uri, nullable: true }
        currentPerformanceAmount: { type: integer, format: int64 }
        currentTierTargetAmount: { type: integer, format: int64, description: 다음 실적 구간의 최소 금액이며, 실적 구간이 없으면 카드 대표 실적 기준 금액 }
        achievementRate: { type: integer, minimum: 0, maximum: 100 }
        currentTier: { type: integer, minimum: 0, description: 현재 실적이 속한 카드 콘텐츠 실적 구간 번호이며, 구간 데이터가 없거나 첫 구간 미만이면 0 }
        nextTier: { type: integer, nullable: true, description: 다음으로 달성할 카드 콘텐츠 실적 구간 번호이며, 최고 구간이면 null }
        isCurrentTierAchieved: { type: boolean }
        remainingAmountToNextTier: { type: integer, format: int64 }
...
    ErrorResponse:
      type: object
      properties:
        success:
          type: boolean
          example: false
        data:
          nullable: true
        error:
          type: object
          properties:
            code:
              type: string
              description: 클라이언트 분기용 오류 코드
              example: AUTHENTICATION_REQUIRED
            message:
              type: string
            fields:
              type: object
              additionalProperties:
                type: string
...